Ran clippy/fmt and moved test to its own file
