+++ polars/polars-lazy/src/tests/mod.rs
@@ -2,6 +2,8 @@
 mod aggregations;
 #[cfg(feature = "test")]
 mod arity;
+#[cfg(feature = "test")]
+mod functions;
 #[cfg(feature = "parquet")]
 mod io;
 #[cfg(feature = "test")]
