+++ polars/tests/it/lazy/queries.rs
@@ -186,22 +186,3 @@ fn test_unknown_supertype_ignore() -> Result<()> {
     assert_eq!(out.shape(), (4, 2));
     Ok(())
 }
-
-#[test]
-#[cfg(feature = "format_str")]
-fn test_format_str() {
-    let a = df![
-        "a" => [1, 2],
-        "b" => ["a", "b"]
-    ].unwrap();
-
-    let out = a.lazy().select([
-        format_str("({}, {}]", [col("a"), col("b")]).alias("formatted") 
-    ]).collect().unwrap();
-
-    let expected = df![
-        "formatted" => ["(1, a]", "(2, b]"]
-    ].unwrap();
-
-    assert!(out.frame_equal_missing(&expected));
-}
